mcp-agent/requirements.txt
# MCP Agent Bridge for LM Studio
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
pydantic>=2.5.0
python-dotenv>=1.0.0
